requirements.txt

Django==2.2.6
djangorestframework==3.10.3
Pillow